% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten_color.R
\name{lighten_color}
\alias{lighten_color}
\title{Lighten a Color}
\usage{
lighten_color(color, factor)
}
\arguments{
\item{color}{A character string representing a color in any valid format (e.g., hex code, color name).}

\item{factor}{A numeric value between 0 and 1 indicating the amount to lighten the color.
A factor of 0 means no change, while a factor of 1 results in white.}
}
\value{
A character string representing the lightened color in RGB format.
}
\description{
This function lightens a given color by a specified factor.
}
